fixed solution of exercise 10 of dox

+++ dox/exercise_10/src/main/java/de/paluno/calculator/client/Client.java
@@ -7,15 +7,17 @@ import org.slf4j.LoggerFactory;
 
 import java.net.MalformedURLException;
 
-public class Client {
+public class Client
+{
 
     private static final Logger LOG = LoggerFactory.getLogger(Client.class);
 
-	/**
-	 * @param args
-	 * @throws MalformedURLException
-	 */
-	public static void main(String[] args) throws MalformedURLException {
+    /**
+     * @param args
+     * @throws MalformedURLException
+     */
+    public static void main(String[] args) throws MalformedURLException
+    {
         CalculationServices calculator = new CalculatorServerService().getCalculatorServerPort();
 
         int add = calculator.add(4, 5);
@@ -24,13 +26,13 @@ public class Client {
         int product = calculator.multiply(4, 5);
         LOG.info(String.format("%12s: %d, %d = %d", "multiply", 4, 5, product));
 
-        int division = calculator.divide(8, 2);
-        LOG.info(String.format("%12s: %d, %d = %d", "division", 8, 2, division));
+        double division = calculator.divide(8, 2);
+        LOG.info(String.format("%12s: %d, %d = %f", "division", 8, 2, division));
 
         int substraction = calculator.substract(0, 1);
         LOG.info(String.format("%12s: %d, %d = %d", "substraction", 0, 1, substraction));
 
         LOG.info(String.format("%12s: %d", "random", calculator.random()));
-	}
+    }
 
 }
